EPMDJ-346. Refactoring for the CoverageController class. Scope is now stored not as a set of ExecData but as a separate structure
