+++ plugins/drill-coverage-plugin/src/adminPartMain/kotlin/Data.kt
@@ -99,3 +99,12 @@ data class TestUsagesInfo(
 
 @Serializable
 data class GatheringState(val state: Boolean)
+
+data class CoverageInfoSet(
+    val associatedTests: List<AssociatedTests>,
+    val coverageBlock: CoverageBlock,
+    val newCoverageBlock: NewCoverageBlock,
+    val newMethodsCoverages: List<SimpleJavaMethodCoverage>,
+    val packageCoverage: List<JavaPackageCoverage>,
+    val testUsages: List<TestUsagesInfo>
+)
